Corrected small const problems in uGene Fixed and added createToken in uGene Added tests for createToken Fixed GFF parsing Fixed loading of features in uGene for GFF/GTF Tested the above. Minor refactor long long int to long int Cleaned up some //TODO

+++ NGS/IO/Writer/uWriterWig.cpp
@@ -22,8 +22,8 @@ void uWriterWig::setValueType(const token_param type){
  	try {
        if  ((token.isParamSet(token_param::CHR))&&(token.isParamSet(token_param::START_POS))&&(token.isParamSet(token_param::END_POS)) )
            {
-                long long int ourstart= utility::stoll(token.getParam(token_param::START_POS));
-                long long int ourend= utility::stoll(token.getParam(token_param::END_POS));
+                long int ourstart= utility::stoll(token.getParam(token_param::START_POS));
+                long int ourend= utility::stoll(token.getParam(token_param::END_POS));
                 int ourspan = ( ourend-ourstart+1);
                 float density =1;
                 if (token.isParamSet(m_scoreType))
